pyproject.toml
[project]
name = "vidarr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "holistictraceanalysis>=0.5.0",
    "lightly-train[timm]>=0.11.4",
    "nvidia-dali-cuda120>=1.51.2",
    "timm>=1.0.21",
    "torchmetrics>=1.8.2",
    "wandb>=0.23.0",
]
